Fixed a reorder multi-edit bug
While looping over the selected rows it seems that the rows change
position resulting in some rows being edited multiple times and others
not being edited at all.

The solution has been to build a list of unique ids for all the selected
rows before editing any rows then iterate through this list, only
editing one row at a time, letting the DataGridView reorder the rows
between each edit if it wishes.  I can't help but think this is a
horribly inefficient way of doing things but it's the only solution I
can think of at present.

Note that there is a horrible lag issue with updating the images of the
pieces in the first column.  At first I thought this was related to the
issue I was working on but it appears not to be.   I have been trying to
find a way to refresh these images but as of yet I haven't be able to.
Note that if you mouse over the image the correct 'floating' image is
being displayed.