Catches a duplicate entry exception for Dictionary
If the same price in the same colour appears in the file the program
would hang.  Catching this exception does allow the program to continue
though as assertion failure is raised in debug mode.

Note, this change is in no way intended to be a fix it is just
highlighting the problem.