Use delegates in "SortSafeSet" deal with toggles
Because toggles are dependant on the current state of the cell a
delegate was employed to allow the new state to be a function of the
current state.  While this has worked for most options I was not able to
calculate the total price column when adding or subtracting a constant
number of items from the quantity so these options still use the
'un-SortSafe' setting method.

I also discovered there already was an ID called "extid" which was a
combination of the part number and colour (which should be sufficient to
uniquely identify rows) so I was able to remove the UID method and just
use the "extid" value instead.