Fixed duplicate row issue
The "extid" column has been defined as the "PrimaryKey" and the
DataTable is now checked for pre-existing rows before adding a new row.

If a pre-existing row is found when adding a new item the currently
existing row is highlighted instead of adding the row.

If duplicate entries are fount when importing from a file the quantities
of the respective rows are summed together and the result displayed as a
single row.